説明画面を変更しました ルーレット削除のタイミングでアラートダイアログ追加 レイアウト修正

+++ app/src/main/java/com/example/biasroulette/RouletteNameAdapter.kt
@@ -1,17 +1,18 @@
 package com.example.biasroulette
 
 import android.annotation.SuppressLint
+import android.content.Intent
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
+import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.RecyclerView
 import io.realm.OrderedRealmCollection
 import io.realm.Realm
 import io.realm.RealmRecyclerViewAdapter
 import io.realm.kotlin.where
-import android.content.Intent
 
 class RouletteNameAdapter (data: OrderedRealmCollection<RouletteName>, ) :
     RealmRecyclerViewAdapter<RouletteName, RouletteNameAdapter.ViewHolder>(data, true) {
@@ -45,21 +46,31 @@ class RouletteNameAdapter (data: OrderedRealmCollection<RouletteName>, ) :
 
         //消去ボタンが押されたときの処理
         holder.roulette_delete_btn.setOnClickListener {
+            val context = holder.itemView.context
 
             //val roulette_name = realm.where<RouletteName>().equalTo("id", roulette_name?.id).findFirst()
 
-            //roulette(項目一つ一つ)の消去
-            val roulettes = realm.where<Roulette>().equalTo("name", roulette_name?.name.toString()).findAll()
-            for (roulette in roulettes) {
-                realm.executeTransaction{
-                    roulette?.deleteFromRealm()
+//            ここにアラートダイアログを入れます
+            AlertDialog.Builder(context)
+                .setMessage("本当に削除しても良いですか？")
+                .setPositiveButton("Yes") { dialog, _ ->
+                    //roulette(項目一つ一つ)の消去
+                    val roulettes = realm.where<Roulette>().equalTo("name", roulette_name?.name.toString()).findAll()
+                    for (roulette in roulettes) {
+                        realm.executeTransaction{
+                            roulette?.deleteFromRealm()
+                        }
+                    }
+                    //roulette_nameの消去
+                    realm.executeTransaction{
+                        roulette_name?.deleteFromRealm()
+                    }
+                    dialog.dismiss()
                 }
-            }
-
-            //roulette_nameの消去
-            realm.executeTransaction{
-                roulette_name?.deleteFromRealm()
-            }
+                .setNegativeButton("No") { dialog, _ ->
+                    dialog.dismiss()
+                }
+                .show()
         }
 
         //編集ボタンが押されたときの処理
